Изменение : Убрали warning C4018 "signed/unsigned mismatch"
git-svn-id: svn://mpc-next.ru/mpc-be/trunk@5518 d191f3aa-8897-45be-b9a4-56e5fb472b67

+++ src/filters/parser/AudioSplitter/AMRFile.cpp
@@ -85,7 +85,7 @@ HRESULT CAMRFile::Open(CBaseSplitterFile* pFile)
 		if (frame.size == 1) {
 			continue;
 		}
-		if (!frame.size || m_pFile->GetRemaining() < frame.size - 1) {
+		if (!frame.size || (UINT64)m_pFile->GetRemaining() < frame.size - 1) {
 			break;
 		}
